Id плейлиста изменено с Long на Integer

+++ src/main/java/com/ilnitsk/animusic/playlist/repository/PlaylistSoundtrackRepository.java
@@ -12,11 +12,11 @@ import org.springframework.transaction.annotation.Transactional;
 public interface PlaylistSoundtrackRepository extends JpaRepository<PlaylistSoundtrack,Long> {
     @Query("SELECT COUNT(s) > 0 FROM PlaylistSoundtrack s where s.playlist.id= :playlist_id and s.soundtrack.id= :soundtrack_id")
     boolean playlistAlreadyContainsSoundtrack(
-            @Param("playlist_id") Long playlist_id, @Param("soundtrack_id") Integer soundtrack_id);
+            @Param("playlist_id") Integer playlist_id, @Param("soundtrack_id") Integer soundtrack_id);
 
     @Modifying
     @Transactional
     @Query("DELETE FROM PlaylistSoundtrack s where s.playlist.id = :playlist_id and s.soundtrack.id = :soundtrack_id")
     void deleteTrackFromPlaylist(
-            @Param("playlist_id") Long playlist_id, @Param("soundtrack_id") Integer soundtrack_id);
+            @Param("playlist_id") Integer playlist_id, @Param("soundtrack_id") Integer soundtrack_id);
 }
